--- a/package.json
+++ b/package.json
@@ -13,6 +13,7 @@
         "clean": "shx rm -rf public/bundle*",
         "lint": "tslint --format stylish --project . 'ts/**/*.ts' 'ts/**/*.tsx'",
         "fix": "tslint --fix --format stylish --project . 'ts/**/*.ts' 'ts/**/*.tsx'",
+        "pre_push": "yarn tsc --noEmit && yarn lint",
         "update:tools": "aws s3 sync --delete s3://docs-markdown/ mdx/tools/",
         "dev": "npm run update:tools; node --max-old-space-size=16384 ./node_modules/webpack-dev-server/bin/webpack-dev-server.js --mode development --content-base public --https",
         "deploy_dogfood": "npm run update:tools; yarn index_docs --environment dogfood; npm run build:prod; aws s3 sync ./public/. s3://dogfood.0xproject.com --profile 0xproject --region us-east-1 --grants read=uri=http://acs.amazonaws.com/groups/global/AllUsers",
@@ -129,6 +130,7 @@
         "css-loader": "0.23.x",
         "extend": "^3.0.2",
         "glob": "^7.1.4",
+        "husky": "^3.0.5",
         "json-stringify-pretty-compact": "^2.0.0",
         "less-loader": "^4.1.0",
         "node-sass": "^4.12.0",
@@ -157,5 +159,10 @@
         "webpack-cli": "3.3.7",
         "webpack-dev-server": "^3.8.0",
         "yargs": "^10.0.3"
+    },
+    "husky": {
+        "hooks": {
+            "pre-push": "yarn pre_push"
+        }
     }
 }